отладка

+++ table.py
@@ -6,6 +6,13 @@ import glob
 from openpyxl.drawing.image import Image
 from openpyxl.styles import Alignment
 
+def get_img(user_id):
+            files = []
+            search_pattern = f"C:/engineer/image/photo_{user_id}_*.jpg"
+            print(f"Ищу файлы по пути: {search_pattern}")
+            files = glob.glob(search_pattern)
+            return files if files else None
+
 def create(user_id, caption=None):
     file_path = None
     data = get_from_db(user_id)['payload']
@@ -22,29 +29,23 @@ def create(user_id, caption=None):
         ws = wb.active
         ws.cell(8, 3, data['Вал сукноведущий'])
 
-        def get_img(user_id):
-            files = []
-            search_pattern = f"C:/engineer/image/photo_{user_id}_*.jpg"
-            print(f"Ищу файлы по пути: {search_pattern}")
-            files = glob.glob(search_pattern)
-            return files if files else None
-        
         photos = get_img(user_id)
         if photos:
             img_col = 67
             for img_path in photos:
                 img = Image(img_path) 
-                img.width = 400
-                img.height = 500
+                img.width = 200
+                img.height = 300
                 anchor = f"B{img_col}"
                 ws.add_image(img, anchor)
-                img_col += 10
+                img_col += 18
+            
         
         print(caption)
         if caption:
-            ws.merge_cells('N38:T38')
-            ws['N38'] = caption
-            ws.cell(row=38, column=14).value = caption 
+            ws.merge_cells('F71:L71')
+            ws['F71'] = caption
+            ws.cell(row=38, column=14).value = caption #убрать это
             ws.cell(row=38, column=14).alignment = openpyxl.styles.Alignment(wrapText=True)
             
         start_row = 14
@@ -75,9 +76,11 @@ def create(user_id, caption=None):
                 os.makedirs(r"c:\engineer\reports", exist_ok=True)
                 file_path = os.path.join('reports', filename)
         wb.save(file_path)
-        for i in photos:
-            if os.path.exists(i):
-                os.remove(i)
+        if photos:
+            for i in photos:
+                if os.path.exists(i):
+                    os.remove(i)
+        
         return file_path
     
 def find_file(user_id):
